Add reflog, add moving most recent commit
Reflog is a great way to quickly find a place you want to go back in time to. Also, sometimes you accidentally committed something to one branch and you want to move it to another branch, but in your panic forget how to move it over. These two concepts can be very useful when combined, and it nice to have a cht.sh reminder of how to accomplish the task.