List executable files found in PATH
Many uses for this!

Say you install a large set of updates, so want to cross-check the old
list to the new one, in order to find out which new executables were
installed.

Or, perhaps less obscure, you might this list, without the `-printf`
part, to check their permission and ownership settings.