Mass-replace tabs with 4-width spacing
This seems to be the predominant choice, and matches the last commit I
just made, so I went ahead and converted them all, and changed any, -
for example, 2-space indents.

Let me know if this is undesired.

To understand why I chose to do this, please refer to the previous
commit's message.