Look instead for '# cheat.sh: ' on any line
This is in regards to a request by Chubin on #134.

I didn't use grep(1), because it wound up being a lot slower than my
method. Relying on tools like that isn't always the best choice, which
is why I often just write in pure-shell, and have done for a couple of
years now. There's a time and a place for those external tools, but I
feel this was not it.

Tested and worked well. I've actually seen no noteworthy slowdowns, -
despite pre-processing the files -- YMMV.